#2204 refactoring
